moondream-requirements.txt
moondream
transformers
torch 
pillow 
einops
pyvips-binary 
pyvips
accelerate